Restructure UI modules
